Move componentCheck component to separate class
